boundary names check added
